+++ src/components/CurrentWeather.jsx
@@ -1,29 +1,15 @@
 
-import React, { useContext } from 'react';
+import React from 'react';
 import { resultadoConsultaContext } from '../context/resultadoConsultaContext';
 
 export const CurrentWeather = () => {
-  const { weatherData } = useContext(resultadoConsultaContext);
+  const { weatherData } =React. useContext(resultadoConsultaContext);
 
-  if (!weatherData) {
-    return <p>Busca una ciudad para ver el clima.</p>;
-  }
   const current = weatherData.list?.[0];
 
-  return (
-    <div className="weather-container">
-            {/* Clima actual */}
-            <div className="weather-header">
-        <img
-          className="weather-icon"
-          src={`https://openweathermap.org/img/wn/${current.weather?.[0]?.icon}@4x.png`}
-          alt={current.weather?.[0]?.description}
-        />
-        <h2>
-          {weatherData.city?.name}, {weatherData.city?.country}
-        </h2>
-      </div>
-      <div className="weather-main">
+    // Tomamos el icono del mediodía si existe, o el primero
+    /*
+<div className="weather-main">
         <div className="weather-main-info">
           <p><strong>Temperatura:</strong> {current.main?.temp} °C</p>
           <p><strong>Sensación térmica:</strong> {current.main?.feels_like} °C</p>
@@ -33,8 +19,30 @@ export const CurrentWeather = () => {
           <p><strong>Descripción:</strong> {current.weather?.[0]?.description}</p>
         </div>
       </div>
-    </div>
+    */
+
+  return (
+
+<div className="weather-container">
+{/* Clima actual */}
+<div className="weather-header">
+  <img
+    className="weather-icon"
+    src={`https://openweathermap.org/img/wn/${current.weather?.[0]?.icon}@4x.png`}
+    alt={current.weather?.[0]?.description}
+  />
+  <div>
+    <h2>
+      {weatherData.city?.name}, {weatherData.city?.country}
+    </h2>
+    <p style={{ fontSize: "1.2rem", margin: "5px 0" }}>
+      {Math.round(current.main?.temp)}°C
+    </p>
+    <p>{current.weather?.[0]?.description}</p>
+  </div>
+</div>
+</div>
   );
-};
+
 
 export default CurrentWeather;
